add test extraction into one file per api resource
